Conflict resolution for data classes
